+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Advanced settings tile to the settings screen
+
 ## [v0.5.8] - 2026-07-17
 
 ### Added