+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.5.9] - 2026-07-18
+
 ### Added
 
 - Bookmarks: add, modify, seek to position
@@ -464,7 +466,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Custom fonts support
 - Dynamic theme support
 
-[Unreleased]: https://github.com/likhithpraveenk/storii/compare/v0.5.8...HEAD
+[Unreleased]: https://github.com/likhithpraveenk/storii/compare/v0.5.9...HEAD
+[v0.5.9]: https://github.com/likhithpraveenk/storii/compare/v0.5.8...v0.5.9
 [v0.5.8]: https://github.com/likhithpraveenk/storii/compare/v0.5.7...v0.5.8
 [v0.5.7]: https://github.com/likhithpraveenk/storii/compare/v0.5.6...v0.5.7
 [v0.5.6]: https://github.com/likhithpraveenk/storii/compare/v0.5.5...v0.5.6